+++ src/main/java/com/hedera/exchange/ERTUtils.java
@@ -55,8 +55,6 @@ package com.hedera.exchange;
 import com.amazonaws.services.kms.AWSKMS;
 import com.amazonaws.services.kms.AWSKMSClientBuilder;
 import com.amazonaws.services.kms.model.DecryptRequest;
-import com.amazonaws.services.sns.AmazonSNS;
-import com.amazonaws.services.sns.AmazonSNSClientBuilder;
 import com.amazonaws.util.Base64;
 import com.hedera.exchange.exchanges.Binance;
 import com.hedera.exchange.exchanges.Bitrex;
@@ -91,9 +89,9 @@ import java.util.Map;
  *
  * @author Anirudh, Cesar
  */
-public final class ExchangeRateUtils {
+public final class ERTUtils {
 
-	private static final Logger LOGGER = LogManager.getLogger(ExchangeRateUtils.class);
+	private static final Logger LOGGER = LogManager.getLogger(ERTUtils.class);
 	private static final Map<String, Class<? extends ExchangeCoin>> EXCHANGES = new HashMap<>();
 	private static final int MILLI_SECS_IN_ONE_HOUR = 3_600_000;
 	private static final int MILLI_SECS_IN_ONE_SEC = 1_000;
@@ -108,7 +106,7 @@ public final class ExchangeRateUtils {
 		EXCHANGES.put("paybito", PayBito.class);
 	}
 
-	private ExchangeRateUtils() {
+	private ERTUtils() {
 		throw new UnsupportedOperationException("Utility class");
 	}
