+++ src/main/java/com/hedera/exchange/ExchangeRateTool.java
@@ -52,7 +52,6 @@ package com.hedera.exchange;
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.hedera.hashgraph.sdk.AccountId;
 import com.hedera.hashgraph.sdk.Client;
 import com.hedera.exchange.database.ExchangeDB;
@@ -66,12 +65,11 @@ import org.apache.logging.log4j.Logger;
 
 import java.io.IOException;
 import java.sql.SQLException;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeoutException;
 
-import static com.hedera.exchange.ExchangeRateUtils.getNodesForClient;
+import static com.hedera.exchange.ERTUtils.getNodesForClient;
 
 
 /**
@@ -134,7 +132,7 @@ public class ExchangeRateTool {
         final Rate currentRate = getCurrentRate(exchangeDB, ertParams);
         final AccountId operatorId = AccountId.fromString(ertParams.getOperatorId());
 
-        final List<Exchange> exchanges = ExchangeRateUtils.generateExchanges(ertParams.getExchangeAPIList());
+        final List<Exchange> exchanges = ERTUtils.generateExchanges(ertParams.getExchangeAPIList());
 
         final ERTProcessLogic proc = new ERTProcessLogic(
                 ertParams.getDefaultHbarEquiv(),
